fix(TUP-21055)Migration check executed at each start of studio https://jira.talendforge.org/browse/TUP-21055
