+++ main/plugins/org.talend.core.repository/src/main/java/org/talend/core/repository/utils/ComponentsJsonModel.java
@@ -1,0 +1,124 @@
+// ============================================================================
+//
+// Copyright (C) 2006-2018 Talend Inc. - www.talend.com
+//
+// This source code is available under agreement available at
+// %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
+//
+// You should have received a copy of the agreement
+// along with this program; if not, write to Talend SA
+// 9 rue Pages 92150 Suresnes, France
+//
+// ============================================================================
+package org.talend.core.repository.utils;
+
+import java.util.List;
+import java.util.Set;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/**
+ * DOC jding  class global comment. Detailled comment
+ */
+public class ComponentsJsonModel {
+
+    @JsonProperty("id")
+    private String id;
+
+    @JsonProperty("version")
+    private String version;
+
+    @JsonProperty("parentId")
+    private String parentId;
+
+    @JsonProperty("configurationType")
+    private String configurationType;
+
+    @JsonProperty("name")
+    private String name;
+
+    @JsonProperty("displayName")
+    private String displayName;
+
+    @JsonProperty("edges")
+    private Set edges;
+
+    @JsonProperty("properties")
+    private List properties;
+
+    @JsonProperty("actions")
+    private Object actions;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getVersion() {
+        return version;
+    }
+
+    public void setVersion(String version) {
+        this.version = version;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+
+    public void setDisplayName(String displayName) {
+        this.displayName = displayName;
+    }
+
+    public String getParentId() {
+        return parentId;
+    }
+
+    public void setParentId(String parentId) {
+        this.parentId = parentId;
+    }
+
+    public String getConfigurationType() {
+        return configurationType;
+    }
+
+    public void setConfigurationType(String configurationType) {
+        this.configurationType = configurationType;
+    }
+
+    public Set getEdges() {
+        return edges;
+    }
+
+    public void setEdges(Set edges) {
+        this.edges = edges;
+    }
+
+    public List getProperties() {
+        return properties;
+    }
+
+    public void setProperties(List properties) {
+        this.properties = properties;
+    }
+
+    public Object getActions() {
+        return actions;
+    }
+
+    public void setActions(Object actions) {
+        this.actions = actions;
+    }
+
+}
